feat: Improve handling of image in product add/edit (to be completed later) #49

+++ products/views.py
@@ -82,6 +82,11 @@ def product_detail(request, product_id):
 @login_required
 def add_product(request):
     """ Sysadmin:  Add a product to the store """
+    # If not a superuser kick user out of function
+    if not request.user.is_superuser:
+        messages.error(request,'Product-Add: restricted to SysAdmin only!')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -108,6 +113,10 @@ def add_product(request):
 @login_required
 def edit_product(request, product_id):
     """ Edit a product in the store """
+    # If not a superuser kick user out of function
+    if not request.user.is_superuser:
+        messages.error(request,'Product-Edit: restricted to SysAdmin only!')
+        return redirect(reverse('home'))
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -136,6 +145,10 @@ def edit_product(request, product_id):
 @login_required
 def delete_product(request, product_id):
     """ Delete a product from the store """
+    # If not a superuser kick user out of function
+    if not request.user.is_superuser:
+        messages.error(request,'Product-Delete: restricted to SysAdmin only!')
+        return redirect(reverse('home'))
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, f'Product { product.sku }, {product.name} deleted!')
